requirements.txt
openai>=1.14.3
python-dotenv>=1.0.0
rich>=13.7.0